dividing into components
