+++ src/components/Overlay.jsx
@@ -1,0 +1,7 @@
+const Overlay = () => {
+    return (
+      <div className="fixed inset-0 bg-gray-900 opacity-50 flex items-center justify-center z-10"></div>
+    );
+  };
+
+  export default Overlay;
